Split basicio.cpp into several files
